pyproject.toml
[tool.poetry]
name = "p6-ds"
version = "0.1.0"
description = ""
authors = ["Alhassane <alhassanne.ahmed@servier.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
seaborn = ">=0.13.1,<0.14.0"
pandas = ">=2.1.4,<2.2.0"
numpy = ">=1.26.3,<1.27.0"
scikit-learn = ">=1.3.2,<1.4.0"
nltk = ">=3.8.1,<3.9.0"
unidecode = ">=1.3.8,<1.4.0"
setuptools = "^69.0.2"
wheel = "^0.42.0"
spacy = ">=3.7.2,<3.8.0"
researchpy = "^0.3.5"
missingno = "^0.5.2"
wordcloud = ">=1.9.3,<1.10.0"
statsmodels = ">=0.14.1,<0.15.0"
gensim = ">=4.3.2,<4.4.0"
transformers = "^4.36.1"
jupyter = "^1.0.0"
python-dotenv = ">=1.0.0,<1.1.0"
click = ">=8.1.7,<8.2.0"
matplotlib = ">=3.8.2,<3.9.0"
tqdm = ">=4.66.1,<4.67.0"
pillow = ">=10.2.0,<10.3.0"
tensorflow-metal = {markers = "sys_platform == \"darwin\" and platform_machine == \"arm64\"", version = "^1.1.0"}
tensorflow = {version = "2.10.0", markers = "platform_machine != \"arm64\""}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
p6-ds = "src.main:main"



